check for image delete

+++ Streetcode/Streetcode.BLL/MediatR/Newss/Update/UpdateNewsHandler.cs
@@ -37,8 +37,10 @@ namespace Streetcode.BLL.MediatR.Newss.Update
             else
             {
                 var img = await _repositoryWrapper.ImageRepository.GetFirstOrDefaultAsync(x => x.Id == response.ImageId);
-                response.Image = null;
-                _repositoryWrapper.ImageRepository.Delete(img);
+                if (img != null)
+                {
+                    _repositoryWrapper.ImageRepository.Delete(img);
+                }
             }
 
             _repositoryWrapper.NewsRepository.Update(news);
